Adjust codestyle, refactor reused methods
