+++ libs/asset-editor/src/lib/components/asset-editor-tabs/asset-editor-contacts/manage-contact-dialog/manage-contact-dialog.component.scss
